Backend Completo

+++ src/user/user.service.ts
@@ -1,8 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from './entities/user.entity';
 import { CreateUserDto } from './dto/create-user.dto';
+import { UpdateUserDto } from './dto/update-user.dto';
+import * as bcrypt from 'bcrypt';
 
 @Injectable()
 export class UserService {
@@ -24,15 +26,30 @@ export class UserService {
   }
 
   async remove(id: number): Promise<void> {
+    const user = await this.findOne(id);
+    if (!user) throw new NotFoundException('Usuário não encontrado');
     await this.usersRepository.delete(id);
   }
 
   create(createUserDto: CreateUserDto): Promise<User> {
-    const user = new User();
-    user.nome = createUserDto.nome;
-    user.email = createUserDto.email;
-    user.password = createUserDto.password;
-    user.role = createUserDto.role;
+    const user = this.usersRepository.create(createUserDto);
+    return this.usersRepository.save(user);
+  }
+
+  async update(id: number, updateUserDto: UpdateUserDto): Promise<User> {
+    const user = await this.usersRepository.preload({
+      id: id,
+      ...updateUserDto,
+    });
+
+    if (!user) {
+      throw new NotFoundException(`Usuário #${id} não encontrado`);
+    }
+
+    if (updateUserDto.password) {
+      user.password = await bcrypt.hash(updateUserDto.password, 10);
+    }
+
     return this.usersRepository.save(user);
   }
 }
